Refactor: Remove edit state from core BoardState object
Editable state was kinda local -- it only lasts as long as the EditPadDialog
stays open, and then doesn't really matter anymore. We're removing the transient
editing state from the global BoardState and keeping it in a pair of variables
specific to the EditPadDialog composable. Kinda forgot about remember{} for
a moment.