Add basic UI tests, upgrade versions of dependencies
Some code has been cleaned up (we didn't require JavaFX Swing or the JPanel
initialization since switching to AudioClip as the thing that we use to play
sounds). Also, we've added UI tests using Jetpack Compose testing.

The tests need to be updated to account for more scenarios, but currently
it's working, at the very least.

Gradle has been updated to 7.4.2.