Refactor: Create observable Board state
The board in its current form is tightly coupled with the UI logic. Rather than
adding persistence to this, I'd like to maintain a separate layer for the UI,
as well as persistence.

The idea for how this is implemented is as follows:
1. `Board` is composed of observable flows, mutable via accessor functions.
2. The UI and the persistence coroutines observe the Board for changes.
	1. UI gets updated on recomposition (should automatically trigger on
	state change).
	2. Persistence triggers and saves the file only on observed state change.

Hooking the board state to the UI and persistence layers isn't done yet.