Add Pad to the serialized datastructure; small refactors
We have added the data from the Pad to the serialization structure of the board.
This is essentially a "complete" state for the serialized data. On further
commits we'll be fixing deserialization to initialize the board state from
the file, as well as updating the UI to make use of this.